:art: Refactor code to flatten data
