+++ src/store/thunks/category.ts
@@ -1,13 +1,8 @@
-import { useSelector } from 'react-redux';
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import { AppDispatch } from '../index';
 import { setData, updateSelectedCategory } from '../slices/category';
-import { getCategoriesByIds, getCategoryData } from '../selectors/category';
-import { getBrandData } from '../selectors/brand';
-import { getModelData } from '../selectors/model';
-import { getVariantsData } from '../selectors/variant';
-import { FlattenedItem, TObject, IActionOptions } from '../../@types/types';
-import { ENTITIES } from '../../constants/entities';
+import { getCategoriesByIds } from '../selectors/category';
+import { TObject, IActionOptions } from '../../@types/types';
 import {
   CATEGORY_SELECTED_UPDATE,
   CATEGORY_SELECTED_ATTEMPT,
@@ -40,7 +35,9 @@ export const attemptHandleUpdateSelectedCategory = createAsyncThunk<
 >(CATEGORY_SELECTED_ATTEMPT, async (requestParams: TObject, thunkAPI) => {
   try {
     const { id, isSelected } = requestParams;
-    const categories = getCategoriesByIds(thunkAPI.getState());
+    const categories: { [key: string]: TObject } = getCategoriesByIds(
+      thunkAPI.getState(),
+    );
     const category = categories?.[id];
     thunkAPI.dispatch(handleUpdateSelectedCategory(requestParams));
     category.brandIds?.forEach((brandId: number) => {
@@ -55,60 +52,3 @@ export const attemptHandleUpdateSelectedCategory = createAsyncThunk<
     return thunkAPI.rejectWithValue(statusText);
   }
 });
-
-export const flattenData = (): FlattenedItem[] => {
-  let uid = 0;
-  const categories = useSelector(getCategoryData);
-  const brands = useSelector(getBrandData);
-  const models = useSelector(getModelData);
-  const variants = useSelector(getVariantsData);
-  const flatData: FlattenedItem[] = [];
-
-  categories?.ids?.forEach((categoryId: number) => {
-    const category = categories?.byId?.[categoryId];
-    flatData.push({
-      ...category,
-      uid: uid++,
-      id: category.id,
-      type: ENTITIES.CATEGORY,
-      name: category.name,
-    });
-
-    category?.brandIds?.forEach((brandId: number) => {
-      const brand = brands?.byId?.[brandId];
-      flatData.push({
-        ...brand,
-        uid: uid++,
-        id: brand.id,
-        type: ENTITIES.BRAND,
-        name: brand?.name,
-        parentId: category?.id,
-      });
-
-      brand?.modelIds?.forEach((modelId: number) => {
-        const model = models?.byId?.[modelId];
-        flatData.push({
-          ...model,
-          uid: uid++,
-          id: model.id,
-          type: ENTITIES.MODEL,
-          name: model?.name,
-          parentId: brand?.id,
-        });
-
-        model?.variantIds?.forEach((variantId: number) => {
-          const variant = variants?.byId?.[variantId];
-          flatData.push({
-            ...variant,
-            uid: uid++,
-            type: ENTITIES.VARIANT,
-            parentId: model?.id,
-            name: variant?.desc,
-          });
-        });
-      });
-    });
-  });
-
-  return flatData;
-};
